fix: add create a new option into flags and goals menu
