+++ ui/dashboard/public/locales/ja/common.json
@@ -210,5 +210,6 @@
   "sort-by": "Sort by: {{sortBy}}",
   "no-options-found": "No options found",
   "create-goal": "Create Goal",
-  "create-a-new-goal": "Create a new Goal"
+  "create-a-new-goal": "Create a new Goal",
+  "create-a-new-flag": "Create a new Flag"
 }
