+++ ui/dashboard/src/pages/experiments/experiments-modal/add-experiment-modal/index.tsx
@@ -17,9 +17,10 @@ import { useTranslation } from 'i18n';
 import { cn } from 'utils/style';
 import { IconInfo, IconPlus } from '@icons';
 import { experimentFormSchema } from 'pages/experiments/form-schema';
+import CreateFlagForm from 'pages/feature-flags/flags-modal/add-flag-modal/create-flag-form';
 import Button from 'components/button';
 import { ButtonBar } from 'components/button-bar';
-import { CreatableSelect } from 'components/creatable-select';
+import { CreatableSelect, CustomOption } from 'components/creatable-select';
 import { ReactDatePicker } from 'components/date-time-picker';
 import Divider from 'components/divider';
 import {
@@ -31,6 +32,7 @@ import {
 import Form from 'components/form';
 import Icon from 'components/icon';
 import Input from 'components/input';
+import DialogModal from 'components/modal/dialog';
 import SlideModal from 'components/modal/slide';
 import TextArea from 'components/textarea';
 import CreateGoalModal from 'elements/create-goal-modal';
@@ -86,6 +88,24 @@ const FeatureFlagStatus = ({
   );
 };
 
+const CreateNewOptionButton = ({
+  text,
+  onClick
+}: {
+  text: string;
+  onClick: () => void;
+}) => (
+  <Button
+    type="button"
+    variant="text"
+    className="h-8 self-center w-full hover:bg-gray-100"
+    onClick={onClick}
+  >
+    <Icon icon={IconPlus} color="primary-500" size={'xs'} />
+    {text}
+  </Button>
+);
+
 const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
   const { t } = useTranslation(['form', 'common']);
   const { notify } = useToast();
@@ -98,6 +118,12 @@ const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
     isOpenCreateGoalModal,
     onOpenCreateGoalModal,
     onHiddenCreateGoalModal
+  ] = useToggleOpen(false);
+
+  const [
+    isOpenCreateFlagModal,
+    onOpenCreateFlagModal,
+    onHiddenCreateFlagModal
   ] = useToggleOpen(false);
 
   const { data: goalCollection, isLoading: isLoadingGoals } = useQueryGoals({
@@ -414,7 +440,7 @@ const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
               control={form.control}
               name={`featureId`}
               render={({ field }) => (
-                <Form.Item className="flex flex-col w-full overflow-hidden">
+                <Form.Item className="flex flex-col w-full">
                   <Form.Label required className="relative w-fit">
                     {t('common:flag')}
                     <Icon
@@ -425,47 +451,64 @@ const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
                     />
                   </Form.Label>
                   <Form.Control>
-                    <DropdownMenu>
-                      <DropdownMenuTrigger
-                        placeholder={t(`experiments.select-flag`)}
-                        label={
-                          featureFlagOptions.find(
-                            item => item.value === field.value
-                          )?.label || ''
-                        }
-                        disabled={isLoadingFeature}
-                        variant="secondary"
-                        className="w-full [&>div>p]:truncate [&>div]:max-w-[calc(100%-36px)]"
-                      />
-                      <DropdownMenuContent
-                        className="w-[502px]"
-                        align="start"
-                        {...field}
-                      >
-                        {featureFlagOptions.map((item, index) => (
-                          <DropdownMenuItem
-                            {...field}
-                            key={index}
-                            value={item.value}
-                            label={item.label}
-                            className="justify-between gap-x-4"
-                            additionalElement={
-                              <FeatureFlagStatus
-                                status={t(
-                                  item.enabled
-                                    ? 'experiments.on'
-                                    : 'experiments.off'
-                                )}
-                                enabled={item.enabled}
+                    <CreatableSelect
+                      value={
+                        featureFlagOptions
+                          .filter(item => item.value === field.value)
+                          ?.map(item => ({
+                            label: item.label,
+                            value: item.value
+                          })) || []
+                      }
+                      isMulti={undefined}
+                      disabled={isLoadingFeature}
+                      loading={isLoadingFeature}
+                      placeholder={t(`experiments.select-flag`)}
+                      options={featureFlagOptions?.map(flag => ({
+                        label: flag.label,
+                        value: flag.value,
+                        enabled: flag.enabled
+                      }))}
+                      onChange={value => field.onChange(value.at(-1)?.value)}
+                      onCreateOption={onOpenCreateFlagModal}
+                      formatCreateLabel={() => (
+                        <CreateNewOptionButton
+                          text={t('common:create-a-new-flag')}
+                          onClick={onOpenCreateFlagModal}
+                        />
+                      )}
+                      noOptionsMessage={() => (
+                        <p>{t('common:no-options-found')}</p>
+                      )}
+                      formatOptionLabel={item => (
+                        <div className="flex items-center h-full w-full justify-between gap-x-2">
+                          <p className="w-full truncate">{item.label}</p>
+                          {!item?.__isNew__ && (
+                            <FeatureFlagStatus
+                              status={t(
+                                item.enabled
+                                  ? 'experiments.on'
+                                  : 'experiments.off'
+                              )}
+                              enabled={item.enabled as boolean}
+                            />
+                          )}
+                        </div>
+                      )}
+                      components={{
+                        Option: props => (
+                          <CustomOption
+                            {...props}
+                            createNewOption={
+                              <CreateNewOptionButton
+                                text={t('common:create-a-new-flag')}
+                                onClick={onOpenCreateFlagModal}
                               />
                             }
-                            onSelectOption={value => {
-                              field.onChange(value);
-                            }}
                           />
-                        ))}
-                      </DropdownMenuContent>
-                    </DropdownMenu>
+                        )
+                      }}
+                    />
                   </Form.Control>
                   <Form.Message />
                 </Form.Item>
@@ -543,42 +586,31 @@ const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
                         label: goal.label,
                         value: goal.value
                       }))}
-                      // const formatOptionLabel = ({ label, enabled }) => {
-                      //   return (
-                      //     <div className="flex justify-between space-x-4 pr-2">
-                      //       <span className="flex-1 truncate">{label}</span>
-                      //       <span
-                      //         className={`border rounded-lg text-sm w-11 flex justify-center ${
-                      //           enabled
-                      //             ? 'bg-primary border-primary text-white'
-                      //             : 'bg-gray-100 border-gray-300'
-                      //         }`}
-                      //       >
-                      //         {enabled ? 'On' : 'Off'}
-                      //       </span>
-                      //     </div>
-                      //   );
-                      // };
                       onChange={value =>
                         field.onChange(value.map(goal => goal.value))
                       }
-                      onCreateOption={onOpenCreateGoalModal}
                       formatCreateLabel={() => (
-                        <Button
-                          variant="text"
-                          className="h-6 self-center w-full"
-                        >
-                          <Icon
-                            icon={IconPlus}
-                            color="primary-500"
-                            size={'xs'}
-                          />
-                          {t('common:create-a-new-goal')}
-                        </Button>
+                        <CreateNewOptionButton
+                          text={t('common:create-a-new-goal')}
+                          onClick={onOpenCreateGoalModal}
+                        />
                       )}
                       noOptionsMessage={() => (
                         <p>{t('common:no-options-found')}</p>
                       )}
+                      components={{
+                        Option: props => (
+                          <CustomOption
+                            {...props}
+                            createNewOption={
+                              <CreateNewOptionButton
+                                text={t('common:create-a-new-goal')}
+                                onClick={onOpenCreateGoalModal}
+                              />
+                            }
+                          />
+                        )
+                      }}
                     />
                   </Form.Control>
                   <Form.Message />
@@ -628,6 +660,21 @@ const AddExperimentModal = ({ isOpen, onClose }: AddExperimentModalProps) => {
           isOpen={isOpenCreateGoalModal}
           onClose={onHiddenCreateGoalModal}
         />
+      )}
+      {isOpenCreateFlagModal && (
+        <DialogModal
+          className="w-[500px] h-full max-h-[700px] overflow-hidden"
+          title={t('common:new-flag')}
+          isOpen={isOpenCreateFlagModal}
+          onClose={onHiddenCreateFlagModal}
+        >
+          <CreateFlagForm
+            className={
+              'flex flex-col flex-1 h-full overflow-auto p-5 pb-[170px]'
+            }
+            onClose={onHiddenCreateFlagModal}
+          />
+        </DialogModal>
       )}
     </SlideModal>
   );
